+++ sw/sensor_board/sensor_board_rev1_test/src/hmi/mod.rs
@@ -1,11 +1,13 @@
+pub mod neopixel;
+
 /// hmi.rs
 /// responsible for all human-machine interfacing (SPI display, buttons, LEDs, etc)
 
 #[allow(unused_imports)]
 use log::{debug, error, info, trace, warn};
 
-use crate::neopixel::{Color, NeoPixel};
 use embassy_time::{Duration, Timer};
+pub use neopixel::{Color, NeoPixel};
 
 pub async fn start_hmi(mut neopixel: NeoPixel<'static>, brightness: u8) {
     let colors = Color::all_colors();
