+++ sw/sensor_board/sensor_board_rev1_test/src/hmi/neopixel.rs
