+++ sw/sensor_board/sensor_board_rev1_test/src/lib.rs
@@ -1,13 +1,13 @@
 #![no_std]
 
-use crate::neopixel::NeoPixel;
+use crate::hmi::neopixel::NeoPixel;
 use core::future::Future;
 use esp_hal::gpio::Output;
 
 pub mod app;
 pub mod asm330;
+pub mod gps;
 pub mod hmi;
-pub mod neopixel;
 pub mod old_asm330;
 
 pub trait BoardPeripherals {
