+++ sw/sensor_board/sensor_board_rev1_test/src/types.rs
@@ -1,0 +1,2 @@
+/// types.rs
+/// Defines all of the common datatypes (IMU/GPS/FUSE/etc)
